refactor: move OrderRow component to containers
